+++ daft/dataframe/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from .dataframe import DataFrame, GroupedDataFrame
+from .dataframe import DataFrame, GroupedDataFrame, dataframe
 
-__all__ = ["DataFrame", "GroupedDataFrame"]
+__all__ = ["DataFrame", "GroupedDataFrame", "dataframe"]
