+++ daft/io/__init__.py
@@ -20,11 +20,14 @@ from daft.io._warc import read_warc
 from daft.io._range import _range
 from daft.io.catalog import DataCatalogTable, DataCatalogType
 from daft.io.file_path import from_glob_path
+from daft.io.source import DataSource, DataSourceTask
 
 __all__ = [
     "AzureConfig",
     "DataCatalogTable",
     "DataCatalogType",
+    "DataSource",
+    "DataSourceTask",
     "GCSConfig",
     "HTTPConfig",
     "IOConfig",
