+++ daft/io/_range.py
@@ -1,20 +1,22 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Callable
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Callable, Iterator, overload
 
 if TYPE_CHECKING:
     from collections.abc import Iterator
 
-from typing import overload
 
 from daft import DataType
 from daft.api_annotations import PublicAPI
-from daft.daft import ScanOperatorHandle
-from daft.dataframe import DataFrame
-from daft.io._generator import GeneratorScanOperator
-from daft.logical.builder import LogicalPlanBuilder
-from daft.logical.schema import Schema
+from daft.dataframe import DataFrame, dataframe
+from daft.io.source import DataSource, DataSourceTask
 from daft.recordbatch.recordbatch import RecordBatch
+from daft.schema import schema
+
+if TYPE_CHECKING:
+    from daft.io.pushdowns import Pushdowns
+    from daft.schema import Schema
 
 
 @overload
@@ -133,11 +135,8 @@ def _range(start: int, end: int | None = None, step: int = 1, partitions: int = 
         start = 0
     else:
         start = start
-    scan_op = RangeScanOperator(start, end, step, partitions)
-    handle = ScanOperatorHandle.from_python_scan_operator(scan_op)
-    builder = LogicalPlanBuilder.from_tabular_scan(scan_operator=handle)
-
-    return DataFrame(builder)
+    source = RangeSource(start, end, step, partitions)
+    return dataframe(source)
 
 
 def _range_generators(
@@ -164,8 +163,65 @@ def _range_generators(
         yield partial(generator, partition_idx)
 
 
-class RangeScanOperator(GeneratorScanOperator):
-    def __init__(self, start: int, end: int, step: int = 1, partitions: int = 1) -> None:
-        schema = Schema._from_field_name_and_types([("id", DataType.int64())])
+# class RangeScanOperator(GeneratorScanOperator):
+#     def __init__(self, start: int, end: int, step: int = 1, partitions: int = 1) -> None:
+#         schema = Schema._from_field_name_and_types([("id", DataType.int64())])
 
-        super().__init__(schema=schema, generators=_range_generators(start, end, step, partitions))
+#         super().__init__(schema=schema, generators=_range_generators(start, end, step, partitions))
+
+
+class RangeSource(DataSource):
+    """RangeSource produces a DataFrame from a range with a given step size."""
+
+    _start: int
+    _end: int
+    _step: int
+    _partitions: int
+
+    def __init__(self, start: int, end: int, step: int = 1, partitions: int = 1) -> None:
+        """Create a RangeSource instance.
+
+        Args:
+            start (int): The start of the range.
+            end (int, optional): The end of the range. If not provided, the start is 0 and the end is `start`.
+            step (int, optional): The step size of the range. Defaults to 1.
+            partitions (int, optional): The number of partitions to split the range into. Defaults to 1.
+        """
+        self._start = start
+        self._end = end
+        self._step = step
+        self._partitions = partitions
+
+    def name(self) -> str:
+        return "RangeSource"
+
+    def schema(self) -> Schema:
+        return schema({"id": DataType.int64()})
+
+    def get_tasks(self, pushdowns: Pushdowns) -> Iterator[DataSourceTask]:
+        step = self._step
+        size = (self._end - self._start) // self._partitions
+        curr_s = self._start
+        curr_e = self._start + size
+        while curr_e <= self._end:
+            yield RangeSourceTask(curr_s, curr_e, step)
+            # update to the next chunk
+            curr_s = curr_e + step
+            curr_e = curr_s + size
+
+
+@dataclass
+class RangeSourceTask(DataSourceTask):
+    _start: int
+    _end: int
+    _step: int
+
+    def schema(self) -> Schema:
+        return schema({"id": DataType.int64()})
+
+    def get_batches(self) -> Iterator[RecordBatch]:
+        import pyarrow as pa
+
+        values = list(range(self._start, self._end, self._step))
+        table = pa.Table.from_arrays([pa.array(values, type=pa.int64())], names=["id"])
+        yield RecordBatch.from_arrow(table)
