--- docs/api/pushdowns.md
@@ -1,31 +1,0 @@
-# Pushdowns
-
-Daft has predicate, projection, and limit pushdowns with expressions being represented by *Terms*. Learn more about [Pushdowns](../advanced/pushdowns.md) in the Daft User Guide.
-
-::: daft.io.pushdowns.Pushdowns
-    options:
-        filters: ["!^_"]
-
-::: daft.io.pushdowns.Term
-    options:
-        filters: ["!^_"]
-
-::: daft.io.pushdowns.Reference
-    options:
-        filters: ["!^_"]
-
-::: daft.io.pushdowns.Literal
-    options:
-        filters: ["!^_"]
-
-::: daft.io.pushdowns.Expr
-    options:
-        filters: ["!^_"]
-
-::: daft.io.pushdowns.Arg
-    options:
-        filters: ["!^_"]
-
-::: daft.io.pushdowns.TermVisitor
-    options:
-        filters: ["!^_"]
